requirements.txt
faiss-cpu
numpy
tqdm
